requirements.txt
bs4==0.0.2
build==1.4.0
DateTime==6.0
mkdocs-git-revision-date-localized-plugin==1.5.1
mkdocs-material==9.7.1
mkdocs-mermaid2-plugin==1.2.3
mkdocs-minify-plugin==0.8.0
mkdocstrings-python==2.0.2
pipdeptree==2.30.0
playwright==1.58.0
PyGithub==2.8.1
pylint==4.0.4
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dotenv==1.2.1
python-semantic-release==10.5.3
requests-mock==1.12.1
twine==6.2.0
